+++ src/main/java/blackjack/domain/participant/Player.java
@@ -1,6 +1,5 @@
 package blackjack.domain.participant;
 
-import blackjack.domain.Money;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Hand;
 import blackjack.domain.card.Score;
